mod:GhRepositoryとGhRepositoryListをそれぞれコンポーネント分割した
