+++ documents/src/components/GhRepository/styles.module.css
@@ -1,0 +1,8 @@
+.previewImage {
+  max-width: 200px;
+  max-height: 100px;
+  object-fit: contain;
+  border-radius: 4px;
+  border: 1px solid #eaeaea;
+  transition: transform 0.2s ease;
+}
