+++ documents/src/components/HomepageUsages/index.js
@@ -1,9 +1,6 @@
 import clsx from "clsx";
-import React, { useEffect, useState } from "react";
 import Translate from "@docusaurus/Translate";
-import { loadFromStorage, setToStorage } from "../../utils/loadFromStorage";
-import { fetchRepositoryData } from "../../utils/fetchRepositoryData";
-import staticRepoData from "../../utils/staticRepoData.json";
+import GhRepository from "@site/src/components/GhRepository/GhRepository";
 
 import styles from "./styles.module.css";
 
@@ -37,7 +34,11 @@ export default function HomepageUsages() {
                 <th>🖼️ Preview </th>
               </tr>
             </thead>
-            <tbody>{OwnerList.map((props) => <GhRepository key={props.owner} {...props} />)}</tbody>
+            <tbody>
+              {OwnerList.map((props) => (
+                <GhRepository key={props.owner} {...props} />
+              ))}
+            </tbody>
           </table>
         </div>
         <div className="container">
@@ -52,100 +53,4 @@ export default function HomepageUsages() {
       </div>
     </section>
   );
-}
-
-function GhRepository({ owner }) {
-  const url = `https://api.github.com/orgs/${owner}/repos`;
-  const cacheKey = `cache_${url}`;
-
-  const [repoData, setRepoData] = useState(null);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState(null);
-
-  const hasStaticData =
-    Array.isArray(staticRepoData) && staticRepoData.length > 0;
-
-  useEffect(() => {
-    // 静的データが存在する場合は、それを使用
-    if (hasStaticData) {
-      setRepoData(staticRepoData);
-      setLoading(false);
-      return;
-    }
-
-    // 静的データがない場合は、ローカルストレージから取得を試みる
-    const cachedData = loadFromStorage(localStorage, cacheKey);
-    if (cachedData) {
-      setRepoData(cachedData);
-      setLoading(false);
-      return;
-    }
-
-    // ローカルストレージにもない場合は、動的にAPIから取得
-    (async () => {
-      const { data, error } = await fetchRepositoryData(url);
-
-      if (error) {
-        setError(error.message);
-      } else {
-        setRepoData(data);
-        setToStorage(localStorage, cacheKey, data);
-      }
-      setLoading(false);
-    })();
-  }, [owner, hasStaticData]);
-
-  if (loading) return;
-  if (error) return;
-  if (!repoData) return;
-
-  return repoData
-    .filter((item) => item.topics.includes("thinklet"))
-    .map((item) => {
-      return GhRepositoryList(item);
-    });
-}
-
-function GhRepositoryList(item) {
-  return (
-    <tr key={item.id || item.node_id || item.full_name}>
-      <td>{item.description}</td>
-      <td>
-        <a href={item.html_url} target="_blank" rel="noopener noreferrer">
-          {item.full_name}
-        </a>
-      </td>
-      <td>{new Date(item.pushed_at).toLocaleDateString()}</td>
-      <td>{convertLicense(item)}</td>
-      <td>
-        <a href={item.html_url} target="_blank" rel="noopener noreferrer">
-          <img
-            src={
-              item.ogImage ||
-              `https://raw.githubusercontent.com/${item.full_name}/${item.defaultBranch}/.github/social-preview.png`
-            }
-            alt={`Preview of ${item.name || item.full_name.split("/")[1]}`}
-            className={styles.previewImage}
-            loading="lazy"
-            onError={(e) => {
-              e.target.onerror = null;
-              e.target.src = `https://opengraph.githubassets.com/1/${item.full_name}`;
-            }}
-          />
-        </a>
-      </td>
-    </tr>
-  );
-}
-
-function convertLicense(item) {
-  if (item.license?.name == null || item.license?.url == null) {
-    return (
-      <a href={item.html_url} target="_blank" rel="noopener noreferrer">
-        See Repository
-      </a>
-    );
-  } else {
-    return item.license.name;
-  }
 }
