+++ documents/src/components/HomepageUsages/styles.module.css
@@ -17,16 +17,3 @@
   border-spacing: 0;
   display: table;
 }
-
-.previewImage {
-  max-width: 200px;
-  max-height: 100px;
-  object-fit: contain;
-  border-radius: 4px;
-  border: 1px solid #eaeaea;
-  transition: transform 0.2s ease;
-}
-
-.previewImage:hover {
-  transform: scale(1.05);
-}
